feat: update middleware structure
