+++ standalone/server/src/loadEnvironment.ts
@@ -1,0 +1,2 @@
+import dotenv from "dotenv"
+dotenv.config()
